pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "transformer-evolution-llm"
version = "0.1.0"
description = "Laptop-friendly evolutionary search for MoE-centric transformer variants."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "transformer-evolution maintainers" }
]
dependencies = [
  "accelerate>=0.33",
  "datasets>=3.0",
  "einops>=0.7",
  "numpy>=1.26",
  "optimum>=1.21",
  "peft>=0.12",
  "pydantic>=2.8",
  "PyYAML>=6.0",
  "rich>=13.7",
  "sentencepiece>=0.2",
  "safetensors>=0.4",
  "tokenizers>=0.19",
  "torch==2.4.0",
  "tqdm>=4.66",
  "transformers>=4.44",
  "typer[all]>=0.12",
  "ujson>=5.9"
]

[project.optional-dependencies]
dev = [
  "bandit>=1.7",
  "black>=24.8",
  "detect-secrets>=1.4",
  "mypy>=1.11",
  "pip-audit>=2.7",
  "pytest>=8.3",
  "pytest-cov>=5.0",
  "ruff>=0.6",
  "types-PyYAML>=6.0.12"
]

[project.scripts]
evo-loop = "transformer_evolution_llm.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
  "E",
  "F",
  "I",
  "N",
  "UP",
  "B",
  "S",
  "C4"
]
ignore = [
  "S101"  # asserts are allowed in tests
]
src = ["src", "tests"]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
warn_return_any = true
warn_unused_ignores = true
strict_optional = true
plugins = ["pydantic.mypy"]
packages = ["transformer_evolution_llm"]

[[tool.mypy.overrides]]
module = ["datasets"]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q --cov=transformer_evolution_llm --cov-report=term-missing"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["transformer_evolution_llm"]

[tool.coverage.report]
show_missing = true
skip_covered = true
